Moved things to separate files to make things more modular and started adding some tests.
