+++ src/AsymPdfIntegrandEvaluator.cpp
@@ -1,0 +1,30 @@
+/***
+ * Copyright (C) 2016 Luca Weihs
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "AsymPdfIntegrandEvaluator.h"
+
+std::complex<double> AsymPdfIntegrandEvaluator::integrand(double x, double t,
+                                                          double maxError) {
+  std::complex<double> val;
+  std::complex<double> i(0, 1);
+  if (t == 0) {
+    val = x;
+  } else {
+    val = asymContCharFunction(t, maxError * t / 2.0) * exp(-i * t * x);
+  }
+  return val / (2.0 * M_PI);
+}
